Extract test utility functions
